+++ kernel_tuner/nvml.py
@@ -20,6 +20,7 @@ class nvml():
         try:
             self._pwr_limit = pynvml.nvmlDeviceGetPowerManagementLimit(self.dev)
             self.pwr_constraints = pynvml.nvmlDeviceGetPowerManagementLimitConstraints(self.dev)
+            self.pwr_limit_default = pynvml.nvmlDeviceGetPowerManagementDefaultLimit(self.dev)
         except pynvml.NVMLError_NotSupported:
             self._pwr_limit = None
             self.pwr_constraints = [1, 0] # inverted range to make all range checks fail
@@ -41,6 +42,10 @@ class nvml():
 
             self.supported_mem_clocks = pynvml.nvmlDeviceGetSupportedMemoryClocks(self.dev)
 
+            self._gr_clock = pynvml.nvmlDeviceGetApplicationsClock(self.dev, pynvml.NVML_CLOCK_GRAPHICS)
+            self._sm_clock = pynvml.nvmlDeviceGetApplicationsClock(self.dev, pynvml.NVML_CLOCK_SM)
+            self._mem_clock = pynvml.nvmlDeviceGetApplicationsClock(self.dev, pynvml.NVML_CLOCK_MEM)
+
             #gather the supported gr clocks for each supported mem clock into a dict
             self.supported_gr_clocks = dict()
             for mem_clock in self.supported_mem_clocks:
@@ -52,6 +57,16 @@ class nvml():
             self.mem_clock_default = None
             self.supported_mem_clocks = []
             self.supported_gr_clocks = dict()
+
+    def __del__(self):
+        #try to restore to defaults
+        if self._pwr_limit != None and self.pwr_limit != self.pwr_limit_default:
+            self.pwr_limit = self.pwr_limit_default
+        if self.gr_clock_default != None:
+            gr_app_clock = pynvml.nvmlDeviceGetApplicationsClock(self.dev, pynvml.NVML_CLOCK_GRAPHICS)
+            mem_app_clock = pynvml.nvmlDeviceGetApplicationsClock(self.dev, pynvml.NVML_CLOCK_MEM)
+            if gr_app_clock != self.gr_clock_default or mem_app_clock != self.mem_clock_default:
+                self.set_clocks(self.mem_clock_default, self.gr_clock_default)
 
     @property
     def pwr_state(self):
@@ -66,7 +81,7 @@ class nvml():
     @pwr_limit.setter
     def pwr_limit(self, new_limit):
         if not self.pwr_constraints[0] <= new_limit <= self.pwr_constraints[1]:
-            raise ValueError("Power limit out of range")
+            raise ValueError(f"Power limit {new_limit} out of range [{self.pwr_constraints[0]}, {self.pwr_constraints[1]}]")
         pynvml.nvmlDeviceSetPowerManagementLimit(self.dev, new_limit)
         self._pwr_limit = pynvml.nvmlDeviceGetPowerManagementLimit(self.dev)
 
@@ -85,7 +100,7 @@ class nvml():
     def set_clocks(self, mem_clock, gr_clock):
         """Set the memory and graphics clock for this device (may require permission)"""
         if not mem_clock in self.supported_mem_clocks:
-            raise ValueError("Illegal value for memmory clock")
+            raise ValueError("Illegal value for memory clock")
         if not gr_clock in self.supported_gr_clocks[mem_clock]:
             raise ValueError("Graphics clock incompatible with memory clock")
         pynvml.nvmlDeviceSetApplicationsClocks(self.dev, mem_clock, gr_clock)
@@ -96,20 +111,20 @@ class nvml():
     @property
     def gr_clock(self):
         """Control the graphics clock (may require permission), only values compatible with the memory clock can be set directly"""
-        return pynvml.nvmlDeviceGetApplicationsClock(self.dev, pynvml.NVML_CLOCK_GRAPHICS)
+        return pynvml.nvmlDeviceGetClockInfo(self.dev, pynvml.NVML_CLOCK_GRAPHICS)
 
     @gr_clock.setter
     def gr_clock(self, new_clock):
-        self.set_clocks(self.mem_clock, new_clock)
+        self.set_clocks(self._mem_clock, new_clock)
 
     @property
     def mem_clock(self):
         """Control the graphics clock (may require permission), only values compatible with the graphics clock can be set directly"""
-        return pynvml.nvmlDeviceGetApplicationsClock(self.dev, pynvml.NVML_CLOCK_MEM)
+        return pynvml.nvmlDeviceGetClockInfo(self.dev, pynvml.NVML_CLOCK_MEM)
 
     @mem_clock.setter
     def mem_clock(self, new_clock):
-        self.set_clocks(new_clock, self.gr_clock)
+        self.set_clocks(new_clock, self._gr_clock)
 
     @property
     def temperature(self):
@@ -145,6 +160,7 @@ class NVMLObserver(BenchmarkObserver):
                 raise ValueError(f"Observable {obs} not in supported: {supported}")
         self.observables = observables
 
+        self.measure_power = False
         needs_power = ["power_readings", "nvml_power", "nvml_energy"]
         if any([obs in needs_power for obs in observables]):
             self.measure_power = True