+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - function to create header file with device targets from stored results
 - support for using tuning results in PythonKernel
 - option to control measurements using observers
+- support for NVML tunable parameters
 
 ## [0.3.2] - 2020-11-04
 ### Added